# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NetMon — Network Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /ssh-hosts:
    get:
      operationId: listSshHosts
      tags: [ssh-hosts]
      summary: List all SSH/Telnet terminal hosts
      responses:
        "200":
          description: List of SSH hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SshHost"
    post:
      operationId: createSshHost
      tags: [ssh-hosts]
      summary: Create an SSH/Telnet terminal host
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SshHostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshHost"
  /ssh-hosts/{id}:
    put:
      operationId: updateSshHost
      tags: [ssh-hosts]
      summary: Update an SSH/Telnet terminal host
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SshHostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshHost"
    delete:
      operationId: deleteSshHost
      tags: [ssh-hosts]
      summary: Delete an SSH/Telnet terminal host
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        canAccessAiTerminal:
          type: boolean
        createdAt:
          type: string
      required: [id, username, role, canAccessAiTerminal, createdAt]
    UserInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        canAccessAiTerminal:
          type: boolean
      required: [username, password]
    UserUpdate:
      type: object
      properties:
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [admin, viewer]
        canAccessAiTerminal:
          type: ["boolean", "null"]
...
    SshHost:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ipAddress:
          type: string
        protocol:
          type: string
          enum: [ssh, telnet]
        port:
          type: integer
        username:
          type: string
        systemType:
          type: string
          enum: [mikrotik, cisco, linux, windows, ubiquiti, fortinet, unknown]
        enabled:
          type: boolean
        orderIndex:
          type: integer
      required: [id, name, ipAddress, protocol, port, username, systemType, enabled, orderIndex]
    SshHostInput:
      type: object
      properties:
        name:
          type: string
        ipAddress:
          type: string
        protocol:
          type: string
          enum: [ssh, telnet]
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        systemType:
          type: string
          enum: [mikrotik, cisco, linux, windows, ubiquiti, fortinet, unknown]
        enabled:
          type: boolean
        orderIndex:
          type: integer
      required: [name, ipAddress, protocol, port, username, password, systemType]
    SshHostUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        protocol:
          type: ["string", "null"]
          enum: [ssh, telnet]
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        systemType:
          type: ["string", "null"]
          enum: [mikrotik, cisco, linux, windows, ubiquiti, fortinet, unknown]
        enabled:
          type: ["boolean", "null"]
        orderIndex:
          type: ["integer", "null"]
...